\name{StoEMMIX-package}
\alias{StoEMMIX-package}
\alias{StoEMMIX}
\docType{package}
\title{
\packageTitle{StoEMMIX}
}
\description{
\packageDescription{StoEMMIX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StoEMMIX}
\packageIndices{StoEMMIX}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{StoEMMIX}

Maintainer: \packageMaintainer{StoEMMIX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
